1. created PO for Burger Memu 2. updated README 3. updated tests for Burger Menu 4. updated Smoke test
