+++ Helper/burgerMenu.js
@@ -1,0 +1,66 @@
+import { expect } from '@playwright/test';
+
+export class BurgerMenu {
+
+    constructor(page) {
+        this.page = page;
+//Header
+this.titlePage = page.locator('span.title');
+
+//Burger Menu
+this.burgerMenu = page.locator('#react-burger-menu-btn');
+this.burgerMenuForm = page.locator('.bm-menu');
+this.allInventary = page.locator('#inventory_sidebar_link');
+this.about = page.locator('#about_sidebar_link');
+this.logout = page.locator('#logout_sidebar_link');
+this.resetAppState = page.locator('#reset_sidebar_link');
+this.burgerMenuCloseButton = page.locator('#react-burger-cross-btn');
+}
+
+//Validate Burger Menu
+async validateBurgerMenu() {
+    await expect(this.burgerMenu).toBeVisible();
+    await this.burgerMenu.click();
+    await expect (this.burgerMenuForm).toBeVisible();
+    await expect(this.allInventary).toBeVisible(); 
+    await expect(this.about).toBeVisible();
+    await expect(this.logout).toBeVisible();
+    await expect(this.resetAppState).toBeVisible(); // unknown functionality only checking that element is present 
+    await expect(this.burgerMenuCloseButton).toBeVisible();
+    await this.burgerMenuCloseButton.click();
+    await expect(this.burgerMenuForm).toBeHidden();
+}
+
+//Validate All Items functionality
+async validateAllItems() {
+    await expect(this.burgerMenu).toBeVisible();
+    await this.burgerMenu.click();
+    await expect(this.allInventary).toBeVisible();
+    await this.allInventary.click();
+    await expect(this.titlePage).toHaveText("Products");
+}
+
+// Validate Logout functionality
+async validateLogout() {
+    await expect(this.burgerMenu).toBeVisible();
+    await this.burgerMenu.click();
+    await expect(this.logout).toBeVisible();
+    await this.logout.click();
+}
+
+// Validate About functionality
+async validateAboutFunctionality() {
+    await expect(this.burgerMenu).toBeVisible();
+    await this.burgerMenu.click();
+    await expect(this.about).toBeVisible();
+    await this.about.click();
+    await this.page.waitForNavigation();
+    const currentUrl = this.page.url();
+    const expectedUrl = 'https://saucelabs.com/';
+    if (currentUrl === expectedUrl) {
+        console.log('The URL is expected:', currentUrl);
+      } else {
+        console.error('The URL does not match the expected URL. The resulting URL is:', currentUrl);
+      }
+    }
+}
