+++ Pages/checkoutCompletePage.js
@@ -15,14 +15,6 @@ this.checkoutTitle = page.locator('.title');
 this.cartIcon = page.locator('#shopping_cart_container');
 this.cartBadge = page.locator('.shopping_cart_badge');
 
-//Burger Menu
-this.burgerMenuForm = page.locator('.bm-menu');
-this.allInventary = page.locator('#inventory_sidebar_link');
-this.about = page.locator('#about_sidebar_link');
-this.logout = page.locator('#logout_sidebar_link');
-this.resetAppState = page.locator('#reset_sidebar_link');
-this.burgerMenuCloseButton = page.locator('#react-burger-cross-btn');
-
 //Confirmation image
 this.confirmationImage =page.locator('img[alt="Pony Express"]');
 
@@ -53,44 +45,6 @@ async validateproductsPageHeader(){
     await expect(this.cartBadge).toBeHidden();
 }
 
-//Validate Burger Menu
-async validateBurgerMenu() {
-    await expect(this.burgerMenu).toBeVisible();
-    await this.burgerMenu.click();
-    await expect (this.burgerMenuForm).toBeVisible();
-    await expect(this.allInventary).toBeVisible(); // unknown functionality only checking that element is present 
-    await expect(this.about).toBeVisible();
-    await expect(this.logout).toBeVisible();
-    await expect(this.resetAppState).toBeVisible(); // unknown functionality only checking that element is present 
-    await expect(this.burgerMenuCloseButton).toBeVisible();
-    await this.burgerMenuCloseButton.click();
-    await expect(this.burgerMenuForm).toBeHidden();
-}
-
-// Validate Logout functionality
-async validateLogout() {
-    await expect(this.burgerMenu).toBeVisible();
-    await this.burgerMenu.click();
-    await expect(this.logout).toBeVisible();
-    await this.logout.click();
-}
-
-// Validate About functionality
-async validateAboutFunctionality() {
-    await expect(this.burgerMenu).toBeVisible();
-    await this.burgerMenu.click();
-    await expect(this.about).toBeVisible();
-    await this.about.click();
-    await this.page.waitForNavigation();
-    const currentUrl = this.page.url();
-    const expectedUrl = 'https://saucelabs.com/';
-    if (currentUrl === expectedUrl) {
-        console.log('The URL is expected:', currentUrl);
-      } else {
-        console.error('The URL does not match the expected URL. The resulting URL is:', currentUrl);
-      }
-}
-
 // Validate checkout complete container
 async validateCheckoutCompleteContainer(){
     await expect(this.confirmationImage).toBeVisible();
@@ -108,7 +62,7 @@ async validateBackToHomeButton() {
 
 // Validate Footer
 async validateFooter() {
-    await expect(this.footerText).toHaveText('© 2024 Sauce Labs. All Rights Reserved. Terms of Service | Privacy Policy');
+    await expect(this.footerText).toHaveText('© 2025 Sauce Labs. All Rights Reserved. Terms of Service | Privacy Policy');
     
     //Check redirection to Twitter page
     await expect(this.twitterLink).toBeVisible();
@@ -117,7 +71,7 @@ async validateFooter() {
         this.twitterLink.click(), 
     ]);
     await newTwitterPage.waitForLoadState();
-    await expect(newTwitterPage).toHaveURL('https://x.com/saucelabs');
+    await expect(newTwitterPage).toHaveURL('https://x.com/saucelabs?mx=2');
     await newTwitterPage.close();
 
     //Check redirection to Facebook page
