+++ Pages/checkoutInformationPage.js
@@ -16,14 +16,6 @@ this.checkoutTitle = page.locator('.title');
 this.cartIcon = page.locator('#shopping_cart_container');
 this.cartBadge = page.locator('.shopping_cart_badge');
 
-//Burger Menu
-this.burgerMenuForm = page.locator('.bm-menu');
-this.allInventary = page.locator('#inventory_sidebar_link');
-this.about = page.locator('#about_sidebar_link');
-this.logout = page.locator('#logout_sidebar_link');
-this.resetAppState = page.locator('#reset_sidebar_link');
-this.burgerMenuCloseButton = page.locator('#react-burger-cross-btn');
-
 //Validate Checkout details form
 this.firstName = page.locator('#first-name');
 this.lastName = page.locator('#last-name');
@@ -63,44 +55,6 @@ async validateCartButton() {
     await expect(this.cartIcon).toBeEnabled();
     console.log("Cart button is visible and enabled");
     await this.cartIcon.click();   
-    }
-
-//Validate Burger Menu
-async validateBurgerMenu() {
-    await expect(this.burgerMenu).toBeVisible();
-    await this.burgerMenu.click();
-    await expect (this.burgerMenuForm).toBeVisible();
-    await expect(this.allInventary).toBeVisible(); // unknown functionality only checking that element is present 
-    await expect(this.about).toBeVisible();
-    await expect(this.logout).toBeVisible();
-    await expect(this.resetAppState).toBeVisible(); // unknown functionality only checking that element is present 
-    await expect(this.burgerMenuCloseButton).toBeVisible();
-    await this.burgerMenuCloseButton.click();
-    await expect(this.burgerMenuForm).toBeHidden();
-    }
-
-// Validate Logout functionality
-async validateLogout() {
-    await expect(this.burgerMenu).toBeVisible();
-    await this.burgerMenu.click();
-    await expect(this.logout).toBeVisible();
-    await this.logout.click();
-    }
-
-// Validate About functionality
-async validateAboutFunctionality() {
-    await expect(this.burgerMenu).toBeVisible();
-    await this.burgerMenu.click();
-    await expect(this.about).toBeVisible();
-    await this.about.click();
-    await this.page.waitForNavigation();
-    const currentUrl = this.page.url();
-    const expectedUrl = 'https://saucelabs.com/';
-    if (currentUrl === expectedUrl) {
-        console.log('The URL is expected:', currentUrl);
-      } else {
-        console.error('The URL does not match the expected URL. The resulting URL is:', currentUrl);
-      }
     }
 
     // Validate Checkout form 
@@ -196,7 +150,7 @@ async validatedCancelButton() {
 
 // Validate Footer
 async validateFooter() {
-    await expect(this.footerText).toHaveText('© 2024 Sauce Labs. All Rights Reserved. Terms of Service | Privacy Policy');
+    await expect(this.footerText).toHaveText('© 2025 Sauce Labs. All Rights Reserved. Terms of Service | Privacy Policy');
     
     //Check redirection to Twitter page
     await expect(this.twitterLink).toBeVisible();
@@ -205,7 +159,7 @@ async validateFooter() {
         this.twitterLink.click(), 
     ]);
     await newTwitterPage.waitForLoadState();
-    await expect(newTwitterPage).toHaveURL('https://x.com/saucelabs');
+    await expect(newTwitterPage).toHaveURL('https://x.com/saucelabs?mx=2');
     await newTwitterPage.close();
 
     //Check redirection to Facebook page
